Fix signed comparison error on Linux.
Add cast to fix "comparison of integer expressions of different
signedness" error.

+++ src/fiends/treasure_table_view.cpp
@@ -58,7 +58,7 @@ TreasureTableView::draw()
         TDrawBuffer b;
         int i = delta.y + y;
         if (size_t(i) < treasureTypes.size()) {
-            auto line = (delta.x < treasureTypes[i].length())
+            auto line = (size_t(delta.x) < treasureTypes[i].length())
                     ? treasureTypes[i].substr(size_t(delta.x))
                     : string();
             if (line.length() < size_t(size.x)) {
